+++ src/constants/keys.ts
@@ -1,4 +1,5 @@
 export enum KeyType {
   ledger = "ledger",
   plaintextKey = "plaintextKey",
+  trezor = "trezor",
 }