+++ src/KeyManager.ts
@@ -1,7 +1,6 @@
 import StellarSdk, { Transaction } from "stellar-sdk";
 
-// TODO:
-// import {albedoHandler} from "./keyTypeHandlers/albedo";
+import { albedoHandler } from "./keyTypeHandlers/albedo";
 import { ledgerHandler } from "./keyTypeHandlers/ledger";
 import { lyraHandler } from "./keyTypeHandlers/lyra";
 import { plaintextKeyHandler } from "./keyTypeHandlers/plaintextKey";
@@ -88,8 +87,7 @@ export class KeyManager {
   constructor(params: KeyManagerParams) {
     this.encrypterMap = {};
     this.keyHandlerMap = {
-      // TODO:
-      // [KeyType.albedo]: albedoHandler,
+      [KeyType.albedo]: albedoHandler,
       [KeyType.ledger]: ledgerHandler,
       [KeyType.lyra]: lyraHandler,
       [KeyType.plaintextKey]: plaintextKeyHandler,