+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [KeyManager] Added Trezor wallet.
+
 ## In master
 
 ## [v0.1.0-rc.1](https://github.com/stellar/js-stellar-wallets/compare/v0.0.9-rc.1...v0.1.0-rc.1)