+++ CHANGELOG.md
@@ -1,8 +1,7 @@
-# Changelog
+## In master
 
 - [KeyManager] Added Trezor wallet.
-
-## In master
+- [KeyManager] Only sign SEP-10 auth transactions with seq number 0.
 
 ## [v0.1.0-rc.1](https://github.com/stellar/js-stellar-wallets/compare/v0.0.9-rc.1...v0.1.0-rc.1)